feat: 결제한 사용자 강의 등록 서비스 추가

+++ src/main/java/nextstep/payments/domain/Payment.java
@@ -35,7 +35,7 @@ public class Payment {
         return nsUserId;
     }
 
-    public Long amount() {
-        return amount;
+    public int amount() {
+        return amount.intValue();
     }
 }
